Refactor: 코드 가독성 향상을 위해 컴포넌트/서비스/타입 파일 분리
