+++ src/types/ApateuGameDataTypes.ts
@@ -1,0 +1,5 @@
+export interface ApateuGameDataTypes {
+  gameMode: string;
+  participants: string[];
+  numberOfFloors: number;
+}
