+++ src/types/FloorNumberAdjustmentButtonsTypes.ts
@@ -1,0 +1,5 @@
+export interface FloorNumberAdjustmentButtonsTypes {
+  numberOfFloors: number;
+  setNumberOfFloors: React.Dispatch<React.SetStateAction<number>>;
+  handleFloorInputChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
+}
