+++ src/types/ParticipantListTypes.ts
@@ -1,0 +1,7 @@
+export interface ParticipantListTypes {
+  participants: string[];
+  handleAddParticipant: React.MouseEventHandler<HTMLButtonElement>;
+  handleRemoveParticipant: (index: number) => void;
+  handleChangeParticipantName: (index: number, name: string) => void;
+  MAX_PARTICIPANTS: number;
+}
