+++ src/types/StartGameButtonTypes.ts
@@ -1,0 +1,4 @@
+export interface StartGameButtonTypes {
+  handleStartGame: () => void;
+  isDisabled: boolean;
+}
